Varios fixes and adjustments. Fixed fatal error on extension activation.
