+++ Extension_AiCrawler_Page_View.php
@@ -19,7 +19,7 @@ $config = Dispatcher::config();
 <p>
 	<?php esc_html_e( 'AI Crawler service is currently', 'w3-total-cache' ); ?>
 	<?php
-	if ( Extension_AiCrawler_Plugin::is_enabled() ) {
+	if ( Extension_AiCrawler_Util::is_enabled() ) {
 		echo ' <span class="w3tc-enabled">' . esc_html__( 'enabled.', 'w3-total-cache' ) . '</span>';
 	} else {
 		echo ' <span class="w3tc-disabled">' . esc_html__( 'disabled.', 'w3-total-cache' ) . '</span>';
