+++ Extensions_Plugin_Admin.php
@@ -55,8 +55,8 @@ class Extensions_Plugin_Admin {
 		add_filter( 'w3tc_admin_menu', array( $this, 'w3tc_admin_menu' ), 10000 );
 		add_action( 'w3tc_settings_page-w3tc_extensions', array( $this, 'w3tc_settings_page_w3tc_extensions' ) );
 
-		if ( Extension_AiCrawler_Plugin::is_allowed() ) {
-			add_filter( 'w3tc_extensions', array( '\W3TC\Extension_AiCrawler_Plugin_Admin', 'w3tc_extensions' ), 10, 2 );
+		if ( Extension_AiCrawler_Util::is_allowed() ) {
+			add_filter( 'w3tc_extensions', array( '\W3TC\Extension_AiCrawler_Plugin_Admin', 'w3tc_extensions' ) );
 		}
 
 		if ( Util_Admin::is_w3tc_admin_page() ) {
